Restrict ir_ad_groups cookie to specific AD groups
Cookies can only be a certain size before the request becomes a
problem on the server or in the browser. For users that are members
of many AD groups, the ir_ad_groups cookie gets large quickly.

We can restrict the groups in this cookie to only those that matter
for access on ir.wsu.edu rather than all of the groups the user
is a member of.